% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_helper}
\alias{grid_helper}
\title{Build a vector from the origin approximating the minimum and maximum values}
\usage{
grid_helper(o, min, max, res)
}
\arguments{
\item{o}{a numeric(1). Origin.}

\item{min}{a numeric(1). Mininum value.}

\item{max}{a numeric(1). Maximum value.}

\item{res}{a numeric(1). Resultuion.}
}
\value{
A numeric.
}
\description{
Build a vector between the minimum and maximum ensuring that the given
origin value is part of the returned vector.
}
