% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_to_raster}
\alias{grid_to_raster}
\title{Rasterize a grid}
\usage{
grid_to_raster(grid_sf, grid_resolution, cname)
}
\arguments{
\item{grid_sf}{an sf object of type polygon.}

\item{grid_resolution}{a numeric of length 1 or 2.}

\item{cname}{a character(1). Name of an attribute in grid_sf.}
}
\value{
a raster (terra object).
}
\description{
Transform a grid (sf object) into a raster (terra object).
}
\seealso{
\link[terra:rasterize]{terra::rasterize} which this function wraps.
}
