% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{is_grid_na}
\alias{is_grid_na}
\title{Are grid attributes NAs?}
\usage{
is_grid_na(x)
}
\arguments{
\item{x}{a grid (sf object).}
}
\value{
a logical
}
\description{
Test whether all the values in the given grid are NA.
}
