% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{make_grid_origin_res}
\alias{make_grid_origin_res}
\title{Build a grid using a known point, a resolution, and a extent}
\usage{
make_grid_origin_res(
  xy_origin,
  xy_min,
  xy_max,
  cell_size,
  crs = 4326,
  id_col = "id"
)
}
\arguments{
\item{xy_origin}{A numeric(2) representing an XY point.}
}
\description{
\code{make_grid_origin_res} returns a grid using the coordinates of an origin
point inside an area and a cell size. The new grid will have a vertex on
the origin and will fit as many cells as possible in the area.
}
