% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropomi.R
\name{read_var}
\alias{read_var}
\title{Read data from a HDF5 file}
\usage{
read_var(v, fname, group = NA)
}
\arguments{
\item{v}{a character(1). Variable's name.}

\item{fname}{a character(1). Path to a HDF5 file.}

\item{group}{a character(1). Name of the variable's group.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ an array.
}\if{html}{\out{</div>}}
}
\description{
DEPRECATED. Read the x variable from the given HDF5 file.
}
