% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropomi.R
\name{tropomi2sf}
\alias{tropomi2sf}
\title{Export from TROPOMI to GeoTIF}
\usage{
tropomi2sf(
  fname,
  vars,
  out_crs = "epsg:4326",
  min_x = -Inf,
  min_y = -Inf,
  max_x = Inf,
  max_y = Inf
)
}
\arguments{
\item{fname}{a character(1). Path to an HDF5 file.}

\item{vars}{a character. Names of the varibles to export. Note that the
first two elements must be the names of the variables longitude and
latitude in that order.}

\item{out_crs}{a spatial reference system for the output tif files.}

\item{min_x, min_y, max_x, max_y}{a numeric(1). Return the points faling in
these ranges.}
}
\value{
a list of sf objects (point).
}
\description{
Export the given HDF5 TROPOMI file as points (sf object).
}
