% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropomi.R
\name{tropomi8day2tif}
\alias{tropomi8day2tif}
\title{Export from TROPOMI to GeoTIF}
\usage{
tropomi8day2tif(fname, vars, out_dir, in_nodata = -999, out_crs = "epsg:4326")
}
\arguments{
\item{fname}{a character(1). Path to an HDF5 file.}

\item{vars}{a character. Names of the varibles to export. Note that the
first two elements must be the names of the variables longitude and
latitude in that order.}

\item{out_dir}{a character(1). Path to a directory.}

\item{in_nodata}{a double(1). The NO DATA flag in the input file.}

\item{out_crs}{a character(1). The spatial reference system for the output
files.}
}
\value{
a character. The path to the tif files.
}
\description{
Export the given HDF5 file to GeoTIF. The HDF5 contains TROPOMI SIF 8 day
gridded observations.
}
