% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropomi.R
\name{tropomiday2grid}
\alias{tropomiday2grid}
\title{Aggregate TROPOMI daily data into a grid}
\usage{
tropomiday2grid(
  file_path,
  vars,
  grid,
  f = "mean",
  in_na_value = 9.96920996838687e+36
)
}
\arguments{
\item{file_path}{a character(1). Path to an HDF5 file.}

\item{vars}{a character. Names of the varibles to export. Note that the
first two elements must be the names of the variables longitude and
latitude in that order.}

\item{grid}{an sf object. A grid used to aggregate the given data.}

\item{f}{a character(1). Name of an aggregation function.}

\item{in_na_value}{a numeric. Numeric value used to represente NAs in the
given file.}
}
\value{
an sf object. The given grid with additional columns.
}
\description{
Assume the give HDF5 file contains points and aggregate them using the given
grid. The HDF5 file contents are assumed to be ungridded TROPOMI SIF daily
observations.
}
\seealso{
Ungridded TROPOMI SIF (at 740nm) \url{https://doi.org/10.22002/D1.1347}
}
