% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{make_grid_origin_res}
\alias{make_grid_origin_res}
\title{Build a grid using a known point, a resolution, and a extent}
\usage{
make_grid_origin_res(
  xy_origin,
  xy_min,
  xy_max,
  cell_size,
  crs = 4326,
  id_col = "id"
)
}
\arguments{
\item{xy_origin}{A numeric(2) representing an XY point.}

\item{xy_min, xy_max}{A pair of numeric vectors with the minimum and maximum
coordiantes of the grid.}

\item{cell_size}{A numeric of length 1 or 2. The width and heigth o a cell.
The units are those of the \code{crs}.}

\item{crs}{Numeric or character representing a Coordinate Reference System.}

\item{id_col}{character of length 1. Name of the column for identifying grid
cells.}
}
\description{
\code{make_grid_origin_res} returns a grid using the coordinates of an origin
point inside an area and a cell size. The new grid will have a vertex on
the origin and will fit as many cells as possible in the area.
}
